chore: apply review suggestions
