+++ eslint-rules/turbo-unfiltered-usage.mjs
@@ -6,8 +6,8 @@
  * repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-const fs = require('fs');
-const path = require('path');
+import fs from 'fs';
+import path from 'path';
 
 /**
  * Reads the root turbo.json configuration and extracts defined targets
@@ -44,43 +44,26 @@ function getTurboTargets() {
 }
 
 /**
- * ESLint rule to prevent circular dependencies and warn about unfiltered turbo usage
+ * ESLint rule to warn about unfiltered turbo usage
  */
-module.exports = {
+export default {
   meta: {
-    type: 'problem',
+    type: 'suggestion',
     docs: {
       description:
-        'Prevent circular dependencies and warn about unfiltered turbo run calls in package.json scripts',
+        'Warn about unfiltered turbo run calls that may have unintended project-wide implications',
       category: 'Best Practices',
       recommended: true,
     },
     fixable: null,
-    schema: [
-      {
-        type: 'object',
-        properties: {
-          allowedDirectTargets: {
-            type: 'array',
-            items: { type: 'string' },
-            description:
-              'Array of turbo targets that are allowed to be called directly',
-          },
-        },
-        additionalProperties: false,
-      },
-    ],
+    schema: [],
     messages: {
-      turboCircular:
-        'Script "{{scriptName}}" calls "turbo run {{target}}" which creates a circular dependency. Turbo will call this script when running "{{target}}".',
       turboUnfiltered:
         'Script "{{scriptName}}" calls "turbo run {{target}}" without a filter, which will run across all packages. Consider using --filter to scope to specific packages or call the local script directly.',
     },
   },
 
   create(context) {
-    const options = context.options[0] || {};
-    const allowedDirectTargets = new Set(options.allowedDirectTargets || []);
     const turboTargets = new Set(getTurboTargets());
 
     return {
@@ -107,63 +90,39 @@ module.exports = {
             return;
           }
 
-          // Check each script for circular dependencies and unfiltered turbo usage
+          // Check each script for unfiltered turbo usage
           Object.entries(packageJson.scripts).forEach(
             ([scriptName, scriptCommand]) => {
               if (typeof scriptCommand !== 'string') {
                 return;
               }
 
-              // Skip if this script name is in the allowed direct targets
-              if (allowedDirectTargets.has(scriptName)) {
-                return;
-              }
-
               const normalizedCommand = scriptCommand.trim();
 
-              // Check for direct turbo calls
+              // Check for unfiltered turbo calls
               const turboRunMatch = normalizedCommand.match(/turbo run (\S+)/);
               if (turboRunMatch) {
                 const turboTarget = turboRunMatch[1];
-                if (turboTargets.has(turboTarget)) {
-                  // Calculate the line number where this script appears
-                  const lines = packageJsonContent.split('\n');
-                  let scriptLine = 1;
-                  let scriptColumn = 1;
+                if (
+                  turboTargets.has(turboTarget) &&
+                  scriptName !== turboTarget
+                ) {
+                  // Skip if it has a filter
+                  if (!normalizedCommand.includes('--filter')) {
+                    // Calculate the line number where this script appears
+                    const lines = packageJsonContent.split('\n');
+                    let scriptLine = 1;
+                    let scriptColumn = 1;
 
-                  for (let i = 0; i < lines.length; i++) {
-                    const line = lines[i];
-                    if (line.includes(`"${scriptName}"`)) {
-                      scriptLine = i + 1;
-                      scriptColumn = line.indexOf(`"${scriptName}"`) + 1;
-                      break;
+                    for (let i = 0; i < lines.length; i++) {
+                      const line = lines[i];
+                      if (line.includes(`"${scriptName}"`)) {
+                        scriptLine = i + 1;
+                        scriptColumn = line.indexOf(`"${scriptName}"`) + 1;
+                        break;
+                      }
                     }
-                  }
 
-                  // Determine if this is a circular dependency or just unfiltered usage
-                  const isCircular = scriptName === turboTarget;
-                  const hasFilter = normalizedCommand.includes('--filter');
-
-                  if (isCircular) {
-                    // Circular dependency - error
-                    context.report({
-                      node,
-                      loc: {
-                        start: { line: scriptLine, column: scriptColumn },
-                        end: {
-                          line: scriptLine,
-                          column: scriptColumn + scriptName.length + 2,
-                        }, // +2 for quotes
-                      },
-                      messageId: 'turboCircular',
-                      data: {
-                        scriptName,
-                        target: turboTarget,
-                      },
-                      severity: 0, // 0 = error
-                    });
-                  } else if (!hasFilter) {
-                    // Unfiltered turbo usage - warning
                     context.report({
                       node,
                       loc: {
@@ -178,7 +137,6 @@ module.exports = {
                         scriptName,
                         target: turboTarget,
                       },
-                      severity: 1, // 1 = warning
                     });
                   }
                 }
