+++ packages/apex-ls-node/package.json
@@ -35,6 +35,5 @@
     "vscode-languageserver": "^9.0.1",
     "vscode-languageserver-textdocument": "^1.0.12",
     "vscode-languageserver-protocol": "^3.17.5"
-  },
-  "devDependencies": {}
+  }
 }
